change in version filed in entities

+++ Framework/MGH.Core.Domain/BaseModels/IVersion.cs
@@ -2,5 +2,5 @@
 
 public interface IVersion
 {
-    long Version { get; set; }
+    int Version { get; set; }
 }
